+++ swift_browser_ui_frontend/src/components/BrowserSecondaryNavbar.vue
@@ -80,18 +80,6 @@
         vertical="center"
         data-testid="decryption-toasts"
       />
-      <c-toasts
-        id="refresh-toasts"
-        data-testid="refresh-toasts"
-      >
-        <p>{{ $t("message.encrypt.enReady") }}</p>
-        <c-button
-          text
-          @click="handleRefreshClick"
-        >
-          {{ $t("message.encrypt.refresh") }}
-        </c-button>
-      </c-toasts>
       <c-spacer />
       <div class="nav-item">
         <c-button
@@ -250,10 +238,6 @@ export default {
           );
         });
       }
-    },
-    handleRefreshClick: function() {
-      document.querySelector("#refresh-toasts").removeToast("refresh-toast");
-      location.reload();
     },
   },
 };