+++ swift_browser_ui_frontend/src/components/UploadProgressBar.vue
@@ -1,12 +1,7 @@
 <template>
   <div class="progress-bar">
     <c-progress-bar
-      v-if="isChunking && !isUploading"
-      hide-details
-    />
-
-    <c-progress-bar
-      v-else-if="isUploading && progress != undefined"
+      v-if="isUploading && progress != undefined"
       :value="(progress * 100).toFixed()"
       single-line
       :label="$t('message.upload.progressLabel')"
@@ -23,20 +18,11 @@
 export default {
   name: "ProgressBar",
   computed: {
-    isChunking() {
-      return this.$store.state.isChunking;
-    },
     isUploading() {
       return this.$store.state.isUploading;
     },
     progress() {
       return this.$store.state.uploadProgress;
-    },
-    encryptedFile () {
-      return this.$store.state.encryptedFile;
-    },
-    fileProgress () {
-      return this.$store.state.encryptedFileProgress;
     },
   },
 };